Unique Constraint required on Driver Name and License No fields. Duplicate records shall not allowed in same Corporate Company.
