+++ RFQ.UI.Infrastructure/Provider/DriverAdaptor.cs
@@ -39,9 +39,9 @@ namespace RFQ.UI.Infrastructure.Provider
                 }
                 return null;
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                throw;
+                throw ex;
             }
         }
